synchronizing updates performed on dev area executables *not* under github control (slac_ts5_dlog.perl, ts5_opts_ts7-1.cfg, ts5_opts_ts7-2.cfg) together with updates performed for offline analysis (slac_ts5_parse_scan.perl & slac_ts5_scan_results.perl). changes include improved robustness against timeouts that apparently occur periodically within the Keyence driver level CCS code, monopolizing the connection to the Keyence controler (if not done, rare collisions apparently occur during array readout (OA command) of latched data values)
these are workarounds that circumvent some deeper issues that seem to remain
at the driver level and/or at the structural level, that are probably unique
to details of the Keyence controller's modes of operation.

slac_ts5_parse_scan.perl now contains lines to analyze, but not use, the
SELFCAL data values that are normally included toward the end of a raft scan.

slac_ts5_scan_results.perl fixes some functionality of offline analysis
that retains the ordering of datasets listed on the commandline (differences
now should reflect <set1> - <set2> for example.